+++ BackEnd/Routes/Rating.Routes.js
@@ -6,5 +6,6 @@ let route = express.Router()
 
 route.post("/addRating", upload.single("productImage"), ratingController.addRating)
 route.get("/getRating/:productId", ratingController.getRatings)
+route.get("/getAllReviews", ratingController.getAllReviews)
 
 module.exports = route